+++ source/common/modules/markdown-editor/theme/main-override.ts
@@ -29,16 +29,7 @@ export const mainOverride = EditorView.baseTheme({
   '.cm-scroller .muted': {
     opacity: '0.2'
   },
-  // KaTeX overrides
-  '.katex': {
-    fontSize: '1.1em', // reduce font-size of math a bit
-    display: 'inline-block', // needed for display math to behave properly
-    userSelect: 'none' // Disable user text selection
-  },
-  '.katex-display, .katex-display > .katex > .katex-html': {
-    width: '100%' // display math should be centered
-  },
-  // Panel button overrides
+  // PANELS
   '.cm-panels .cm-button': {
     backgroundImage: 'none',
     backgroundColor: 'inherit',
@@ -55,63 +46,15 @@ export const mainOverride = EditorView.baseTheme({
   '.cm-panel.cm-search': {
     userSelect: 'none' // prevent search panel text elements from being selected
   },
-  // Tooltips
+  // TOOLTIPS
   '.cm-tooltip': {
     padding: '4px'
   },
-  // File preview tooltips
-  '.editor-note-preview': {
-    maxWidth: '300px',
-    padding: '5px',
-    fontSize: '100%'
-  },
-  '.editor-note-preview h1': { fontSize: '140%' },
-  '.editor-note-preview h2': { fontSize: '130%' },
-  '.editor-note-preview h3': { fontSize: '120%' },
-  '.editor-note-preview h4': { fontSize: '110%' },
-  '.editor-note-preview h5': { fontSize: '100%' },
-  '.editor-note-preview h6': { fontSize: '100%' },
-  '.editor-note-preview .note-content': { margin: '10px 0' },
-  '.editor-note-preview .metadata': {
-    color: 'rgb(200, 200, 200)',
-    fontSize: '80%'
-  },
-  '.editor-note-preview .actions': {
-    margin: '5px 0'
-  },
   // Footnotes
-  '.footnote': {
+  '.footnote, .footnote-ref-label': {
     verticalAlign: 'super',
     fontSize: '80%'
-  },
-  '.footnote-ref-label': {
-    verticalAlign: 'super',
-    fontSize: '80%'
-  },
-  // Define the readability classes. Red, orange, and yellow indicate bad scores
-  // Purple and blue indicate average scores, and green indicates good scores
-  '&light .cm-readability-0': { backgroundColor: '#ff0000aa', color: '#444' },
-  '&light .cm-readability-1': { backgroundColor: '#f67b2baa', color: '#444' },
-  '&light .cm-readability-2': { backgroundColor: '#e5a14faa', color: '#444' },
-  '&light .cm-readability-3': { backgroundColor: '#e3e532aa', color: '#444' },
-  '&light .cm-readability-4': { backgroundColor: '#d4c1fdaa', color: '#444' },
-  '&light .cm-readability-5': { backgroundColor: '#538fe9aa', color: '#444' },
-  '&light .cm-readability-6': { backgroundColor: '#53bce9aa', color: '#444' },
-  '&light .cm-readability-7': { backgroundColor: '#53e7e9aa', color: '#444' },
-  '&light .cm-readability-8': { backgroundColor: '#4ad14caa', color: '#444' },
-  '&light .cm-readability-9': { backgroundColor: '#53e955aa', color: '#444' },
-  '&light .cm-readability-10': { backgroundColor: '#7cf87eaa', color: '#444' },
-  '&dark .cm-readability-0': { backgroundColor: '#ff0000aa', color: '#ccc' },
-  '&dark .cm-readability-1': { backgroundColor: '#f67b2baa', color: '#ccc' },
-  '&dark .cm-readability-2': { backgroundColor: '#e5a14faa', color: '#ccc' },
-  '&dark .cm-readability-3': { backgroundColor: '#e3e532aa', color: '#ccc' },
-  '&dark .cm-readability-4': { backgroundColor: '#d4c1fdaa', color: '#ccc' },
-  '&dark .cm-readability-5': { backgroundColor: '#538fe9aa', color: '#ccc' },
-  '&dark .cm-readability-6': { backgroundColor: '#53bce9aa', color: '#ccc' },
-  '&dark .cm-readability-7': { backgroundColor: '#53e7e9aa', color: '#ccc' },
-  '&dark .cm-readability-8': { backgroundColor: '#4ad14caa', color: '#ccc' },
-  '&dark .cm-readability-9': { backgroundColor: '#53e955aa', color: '#ccc' },
-  '&dark .cm-readability-10': { backgroundColor: '#7cf87eaa', color: '#ccc' }
+  }
 })
 
 export const defaultLight = EditorView.theme({}, { dark: false })