+++ CHANGELOG.md
@@ -24,6 +24,8 @@
 - `fsal-directory::removeChild` now calls `pathExists` instead of `isFile` to
   make sure directories also are removed in removeChild
 - `pathExists` wraps `fs.promises.access`.
+- Move plugin-specific base styles from the main override to the respective
+  plugin files
 
 # 3.0.2